Initial attempt at splitting off the LMOD hooks into a seperate SitePackage.lua file
