+++ EESSI-install-software.sh
@@ -248,11 +248,20 @@ fi
 ### add packages here
 
 echo ">> Creating/updating Lmod cache..."
-export LMOD_RC="${EASYBUILD_INSTALLPATH}/.lmod/lmodrc.lua"
+export LMOD_CONFIG_DIR="${EASYBUILD_INSTALLPATH}/.lmod"
+lmod_rc_file="$LMOD_CONFIG_DIR/lmodrc.lua"
 lmodrc_changed=$(cat ${pr_diff} | grep '^+++' | cut -f2 -d' ' | sed 's@^[a-z]/@@g' | grep '^create_lmodrc.py$' > /dev/null; echo $?)
-if [ ! -f $LMOD_RC ] || [ ${lmodrc_changed} == '0' ]; then
+if [ ! -f $lmod_rc_file ] || [ ${lmodrc_changed} == '0' ]; then
     python3 $TOPDIR/create_lmodrc.py ${EASYBUILD_INSTALLPATH}
-    check_exit_code $? "$LMOD_RC created" "Failed to create $LMOD_RC"
+    check_exit_code $? "$lmod_rc_file created" "Failed to create $lmod_rc_file"
+fi
+
+echo ">> Creating/updating Lmod SitePackage.lua ..."
+export LMOD_PACKAGE_PATH=${EASYBUILD_INSTALLPATH}/.lmod"
+lmod_sitepackage_file="$LMOD_PACKAGE_PATH/SitePackage.lua"
+if [ ! -f $lmod_sitepackage_file ] || [ ${sitepackage_changed} == '0' ]; then
+    python3 $TOPDIR/create_lmodsitepackage.py ${EASYBUILD_INSTALLPATH}
+    check_exit_code $? "$lmod_sitepackage_file created" "Failed to create $lmod_sitepackage_file"
 fi
 
 $TOPDIR/update_lmod_cache.sh ${EPREFIX} ${EASYBUILD_INSTALLPATH}
